Fixed minor issues, migrate and migrate rollback working

+++ database/migrations/2024_07_25_223141_create_catalog_categories_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('catalog_categories', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('taxonomy_id');
+            $table->unsignedBigInteger('taxonomy_id');
             $table->string('name');
             $table->text('description')->nullable();
             $table->nestedSet();
